% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligand_receptor.R
\name{analyze_LR}
\alias{analyze_LR}
\title{Generate ligand-receptor table based on gene signatures}
\usage{
analyze_LR(signatures, db)
}
\arguments{
\item{signatures}{Output of Seurat function FindAllMarkers() with only.pos = T}

\item{db_path}{Path to static CellTalk database as .tsv}
}
\value{
A LR table with labeled interactions
}
\description{
Generate ligand-receptor table based on gene signatures
}
