% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_functions_fixed.R
\name{create_enhanced_vlnplot}
\alias{create_enhanced_vlnplot}
\title{Enhanced VlnPlot with custom layout}
\usage{
create_enhanced_vlnplot(
  obj,
  features,
  group.by = "samples",
  pt.size = 0,
  ncol = 3,
  theme_choice = "classic"
)
}
\description{
Enhanced VlnPlot with custom layout
}
