% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_interactive_plots.R
\name{create_linked_plots}
\alias{create_linked_plots}
\title{Add cross-plot linking functionality}
\usage{
create_linked_plots(plot_outputs)
}
\arguments{
\item{plot_outputs}{List of plotly outputs to link}
}
\description{
Add cross-plot linking functionality
}
