% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_interactive_plots.R
\name{create_selection_handlers}
\alias{create_selection_handlers}
\title{Create selection event handlers for interactive plots}
\usage{
create_selection_handlers(input, output, obj)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{obj}{Seurat object}
}
\description{
Create selection event handlers for interactive plots
}
