% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligand_receptor.R
\name{crossreference_degs}
\alias{crossreference_degs}
\title{Find DE'ed ligand-receptor pairs in LR table}
\usage{
crossreference_degs(lr_table, deg_table, db)
}
\arguments{
\item{lr_table}{Dataframe generated from analyze_LR()}

\item{deg_table}{Dataframe generated from find_all_degs()}

\item{db_path}{Path to static CellTalk database as .tsv}
}
\value{
A LR table with annotated DEGs
}
\description{
Find DE'ed ligand-receptor pairs in LR table
}
