% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{explorer_server}
\alias{explorer_server}
\title{server side functions related to \code{explorer_sidebar_ui}}
\usage{
explorer_server(input, output, session, data, verbose = FALSE)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}

\item{data}{the Seurat object and related parameters}

\item{verbose}{for debug use}
}
\value{
server side functions related to \code{explorer_sidebar_ui}
}
\description{
server side functions related to \code{explorer_sidebar_ui}
}
