% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{explorer_sidebar_ui}
\alias{explorer_sidebar_ui}
\title{some menu items of the dashboard sidebar}
\usage{
explorer_sidebar_ui()
}
\value{
return some menu items for the dashboard
}
\description{
to generate some menu items for the dashboard, which can be integrated to other
packages, such as 'fentouxungui/SeuratExplorerServer' from github.
}
\examples{
explorer_sidebar_ui()

}
