% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_functions_fixed.R
\name{feature_selection_server}
\alias{feature_selection_server}
\title{Feature Selection Server Module}
\usage{
feature_selection_server(id, selected_features_reactive)
}
\description{
Feature Selection Server Module
}
