% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{flexible_vln}
\alias{flexible_vln}
\title{VlnPlot wrapper for stacked/non-stacked plots}
\usage{
flexible_vln(obj, features, split.by = NULL, idents = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{Vector of gene names}

\item{split.by}{Metadata column name to split violin}

\item{idents}{Vector of identies to plot}
}
\description{
VlnPlot wrapper for stacked/non-stacked plots
}
