% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligand_receptor.R
\name{get_all_degs}
\alias{get_all_degs}
\title{Wrapper for FindMarkers() with formating for downstream LR analysis}
\usage{
get_all_degs(
  obj,
  annotation,
  condition,
  reference,
  experimental,
  padj.cut = 0.05,
  log2fc.cut = 0.1
)
}
\arguments{
\item{obj}{Seurat object}

\item{annotation}{Column name of object meta.data to annotate identities by}

\item{condition}{Column name of object meta.data to test DE against}

\item{reference}{Value of condition column to use as reference for DE}

\item{experimental}{Value of condition column to use as experimental condition for DE}

\item{padj.cut}{Threshold for adjusted p-value}

\item{log2fc.cut}{Threshold for log2-fold-change}
}
\value{
A single, merged DF of DEGs across every identity
}
\description{
Wrapper for FindMarkers() with formating for downstream LR analysis
}
