% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interactive_3d_dimplot}
\alias{interactive_3d_dimplot}
\title{Create a 3D interactive plot from Seurat dimensional reduction}
\usage{
interactive_3d_dimplot(
  obj,
  reduction = "umap",
  group.by = "seurat_clusters",
  pt.size = 1
)
}
\arguments{
\item{obj}{Seurat object}

\item{reduction}{Dimensional reduction to use (must have 3 dimensions)}

\item{group.by}{Grouping variable for coloring}

\item{pt.size}{Point size}
}
\value{
A plotly 3D scatter plot
}
\description{
Create a 3D interactive plot from Seurat dimensional reduction
}
