% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_functions_fixed.R
\name{interactive_complex_heatmap}
\alias{interactive_complex_heatmap}
\title{Interactive Complex Heatmap with viridis colors}
\usage{
interactive_complex_heatmap(
  obj,
  features,
  group.by = "seurat_clusters",
  color_palette = "viridis",
  cluster_columns = TRUE,
  cluster_rows = TRUE
)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{Features to plot}

\item{group.by}{Grouping variable}

\item{color_palette}{Color palette choice}
}
\description{
Interactive Complex Heatmap with viridis colors
}
