% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interactive_dimplot}
\alias{interactive_dimplot}
\title{Convert a Seurat DimPlot to an interactive plotly object}
\usage{
interactive_dimplot(
  p,
  obj,
  reduction = "umap",
  group.by = "seurat_clusters",
  height = NULL
)
}
\arguments{
\item{p}{A ggplot object from Seurat::DimPlot}

\item{obj}{The Seurat object}

\item{reduction}{The reduction to use}

\item{group.by}{The metadata column to group by}

\item{height}{Height in pixels for the plot}
}
\value{
A plotly object
}
\description{
Convert a Seurat DimPlot to an interactive plotly object
}
