% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_functions_fixed.R
\name{interactive_dimplot_fixed}
\alias{interactive_dimplot_fixed}
\title{Fixed interactive DimPlot with proper tooltips and height control}
\usage{
interactive_dimplot_fixed(
  p,
  obj,
  reduction = "umap",
  group.by = "seurat_clusters",
  hw_ratio = 0.9,
  width_px = 600,
  tooltip_vars = c("nFeature_RNA", "nCount_RNA"),
  theme_choice = "plotly_white"
)
}
\arguments{
\item{p}{ggplot object from Seurat::DimPlot}

\item{obj}{Seurat object}

\item{reduction}{Reduction type}

\item{group.by}{Grouping column}

\item{hw_ratio}{Height/width ratio}

\item{width_px}{Width in pixels}

\item{tooltip_vars}{Variables to show in tooltip}

\item{theme_choice}{Theme for the plot}
}
\description{
Fixed interactive DimPlot with proper tooltips and height control
}
