% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interactive_dotplot}
\alias{interactive_dotplot}
\title{Convert a Seurat DotPlot to an interactive plotly object}
\usage{
interactive_dotplot(
  p,
  obj,
  features,
  group.by = "seurat_clusters",
  height = NULL
)
}
\arguments{
\item{p}{A ggplot object from Seurat::DotPlot}

\item{obj}{The Seurat object}

\item{features}{The features being plotted}

\item{group.by}{The metadata column to group by}
}
\value{
A plotly object
}
\description{
Convert a Seurat DotPlot to an interactive plotly object
}
