% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interactive_empty_plot}
\alias{interactive_empty_plot}
\title{Create an interactive empty plot for error cases}
\usage{
interactive_empty_plot(message = "Please input correct features!")
}
\arguments{
\item{message}{The error message to display}
}
\value{
A plotly object
}
\description{
Create an interactive empty plot for error cases
}
