% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interactive_featureplot}
\alias{interactive_featureplot}
\title{Convert a Seurat FeaturePlot to an interactive plotly object}
\usage{
interactive_featureplot(
  p,
  obj,
  features,
  reduction = "umap",
  slot = "data",
  height = NULL,
  hover_info = "all",
  drag_mode = "pan"
)
}
\arguments{
\item{p}{A ggplot object from Seurat::FeaturePlot}

\item{obj}{The Seurat object}

\item{features}{The features being plotted}

\item{reduction}{The reduction to use}

\item{slot}{The data slot used}

\item{height}{Height in pixels for the plot}

\item{hover_info}{Information to show on hover}

\item{drag_mode}{Plotly drag mode}
}
\value{
A plotly object
}
\description{
Convert a Seurat FeaturePlot to an interactive plotly object
}
