% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_interactive_plots.R
\name{interactive_featureplot_enhanced}
\alias{interactive_featureplot_enhanced}
\title{Enhanced interactive FeaturePlot with expression-aware tooltips}
\usage{
interactive_featureplot_enhanced(
  obj,
  features,
  reduction = "umap",
  slot = "data",
  assay = "RNA",
  downsample = 10000,
  height = 500
)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{Features to plot}

\item{reduction}{Dimensional reduction}

\item{slot}{Data slot to use}

\item{assay}{Assay to use}
}
\description{
Enhanced interactive FeaturePlot with expression-aware tooltips
}
