% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interactive_vlnplot}
\alias{interactive_vlnplot}
\title{Convert a Seurat VlnPlot to an interactive plotly object}
\usage{
interactive_vlnplot(
  p,
  obj,
  features,
  group.by = "seurat_clusters",
  slot = "data",
  height = NULL
)
}
\arguments{
\item{p}{A ggplot object from Seurat::VlnPlot}

\item{obj}{The Seurat object}

\item{features}{The features being plotted}

\item{group.by}{The metadata column to group by}

\item{slot}{The data slot used}
}
\value{
A plotly object
}
\description{
Convert a Seurat VlnPlot to an interactive plotly object
}
