% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_functions_fixed.R
\name{interactive_vlnplot_fixed}
\alias{interactive_vlnplot_fixed}
\title{Interactive VlnPlot with proper height control}
\usage{
interactive_vlnplot_fixed(
  p,
  obj,
  features,
  group.by = "seurat_clusters",
  hw_ratio = 0.9,
  width_px = 600,
  theme_choice = "plotly_white"
)
}
\description{
Interactive VlnPlot with proper height control
}
