% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{legit_annotations}
\alias{legit_annotations}
\title{Get legit annotations from metadata column names}
\usage{
legit_annotations(df, max_unique = 100)
}
\arguments{
\item{df}{Metadata data.frame from Seurat object}

\item{max_unique}{Cut-off for total unique elements in annotations}
}
\description{
Get legit annotations from metadata column names
}
