% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_interactive_plots.R
\name{optimize_for_performance}
\alias{optimize_for_performance}
\title{Performance-optimized plotting for large datasets}
\usage{
optimize_for_performance(
  obj,
  max_cells = 10000,
  sampling_method = "stratified"
)
}
\arguments{
\item{obj}{Seurat object}

\item{max_cells}{Maximum cells to display}

\item{sampling_method}{Method for downsampling ('random', 'stratified')}
}
\description{
Performance-optimized plotting for large datasets
}
