% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_heatmap.R
\name{plotly_heatmap}
\alias{plotly_heatmap}
\title{Create an interactive plotly heatmap from Seurat data}
\usage{
plotly_heatmap(
  obj,
  features,
  group.by = "seurat_clusters",
  assay = "RNA",
  slot = "data",
  downsample = 1000,
  color_palette = "default"
)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{Features to plot}

\item{group.by}{Grouping variable}

\item{assay}{Assay to use}

\item{slot}{Data slot to use}

\item{downsample}{Number of cells to downsample to}

\item{color_palette}{Color palette option}
}
\value{
A plotly heatmap object
}
\description{
Create an interactive plotly heatmap from Seurat data
}
