% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_gene_annotations}
\alias{prepare_gene_annotations}
\title{extract the features from row names or annotation slot of assays}
\usage{
prepare_gene_annotations(obj, verbose = FALSE)
}
\arguments{
\item{obj}{a Seurat object}

\item{verbose}{whether output messages}
}
\value{
a list with data.frames
}
\description{
extract the features from row names or annotation slot of assays
}
\examples{
#NULL
}
