% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_seurat_object}
\alias{prepare_seurat_object}
\title{prepare_seurat_object}
\usage{
prepare_seurat_object(obj, verbose = FALSE)
}
\arguments{
\item{obj}{Seurat Object}

\item{verbose}{if TRUE, output messages.}
}
\value{
A updated Seurat Object
}
\description{
prepare_seurat_object
}
