% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{read_genes}
\alias{read_genes}
\title{Parse input of comma seperated genes}
\usage{
read_genes(str, obj)
}
\arguments{
\item{str}{string to parse}

\item{obj}{Seurat object to crossreference parsed genes}
}
\description{
Parse input of comma seperated genes
}
