% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{replace_nones}
\alias{replace_nones}
\title{Replace 'None' with NULL for meaningful selections}
\usage{
replace_nones(input)
}
\arguments{
\item{input}{string}
}
\description{
Replace 'None' with NULL for meaningful selections
}
