% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot_fixes.R
\name{server_vlnplot_enhanced}
\alias{server_vlnplot_enhanced}
\title{Fixed server function for interactive violin plots}
\usage{
server_vlnplot_enhanced()
}
\description{
Fixed server function for interactive violin plots
}
