% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cellRatioPlot}
\alias{cellRatioPlot}
\title{绘制细胞比例柱状图}
\usage{
cellRatioPlot(
  object = NULL,
  sample.name = NULL,
  sample.order = NULL,
  celltype.name = NULL,
  celltype.order = NULL,
  facet.name = NULL,
  facet.order = NULL,
  col.width = 0.7,
  flow.alpha = 0.25,
  flow.curve = 0,
  color.choice = NULL
)
}
\arguments{
\item{object}{an seurat object}

\item{sample.name}{x axis}

\item{sample.order}{order for x axis}

\item{celltype.name}{column fill by}

\item{celltype.order}{order for fill by}

\item{facet.name}{分面所有的列}

\item{facet.order}{分面的order}

\item{col.width}{column width, from 0-1}

\item{flow.alpha}{中间连接线的透明度}

\item{flow.curve}{中间连接线的弯曲度}

\item{color.choice}{fill by所用的颜色系列}
}
\value{
a ggplot2 object
}
\description{
绘制细胞比例柱状图，支持分面，源代码参考了: https://github.com/junjunlab/scRNAtoolVis/blob/master/R/cellRatioPlot.R
并做了修改。
}
\examples{
#NULL
}
