% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cellRatioPlot}
\alias{cellRatioPlot}
\title{plot cell percentage barplot}
\usage{
cellRatioPlot(
  object = NULL,
  sample.name = NULL,
  sample.order = NULL,
  celltype.name = NULL,
  celltype.order = NULL,
  facet.name = NULL,
  facet.order = NULL,
  col.width = 0.7,
  flow.alpha = 0.25,
  flow.curve = 0,
  color.choice = NULL
)
}
\arguments{
\item{object}{an Seurat object}

\item{sample.name}{x axis}

\item{sample.order}{order for x axis}

\item{celltype.name}{column fill by}

\item{celltype.order}{order for fill by}

\item{facet.name}{column name for facet}

\item{facet.order}{the order for facet}

\item{col.width}{column width, from 0-1}

\item{flow.alpha}{transparency for flow}

\item{flow.curve}{curve for flow}

\item{color.choice}{color choice for fill}
}
\value{
a ggplot2 object
}
\description{
support facet, codes refer to: https://github.com/junjunlab/scRNAtoolVis/blob/master/R/cellRatioPlot.R, with modification
}
\examples{
#NULL
}
