% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{explorer_body_ui}
\alias{explorer_body_ui}
\title{body - ui functions for Seurat explorer functions}
\usage{
explorer_body_ui(tab_list)
}
\arguments{
\item{tab_list}{settings for each seurat function}
}
\description{
body - ui functions for Seurat explorer functions
}
