% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{explorer_server}
\alias{explorer_server}
\title{server functions for Seurat Explorer menuitems
R shiny server side for Seurat plot and DEGs analysis functions}
\usage{
explorer_server(input, output, session, data)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}

\item{data}{the Seurat object and related parameters}
}
\description{
server functions for Seurat Explorer menuitems
R shiny server side for Seurat plot and DEGs analysis functions
}
