% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{explorer_sidebar_ui}
\alias{explorer_sidebar_ui}
\title{sidebar - ui functions for Seurat explorer functions}
\usage{
explorer_sidebar_ui()
}
\description{
sidebar - ui functions for Seurat explorer functions
}
