% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getColors}
\alias{getColors}
\title{getColors}
\usage{
getColors(color.platte = NULL, choice = "default", n = NULL)
}
\arguments{
\item{color.platte}{predefined color list}

\item{choice}{color name}

\item{n}{how many colors to return}
}
\value{
a color list
}
\description{
getColors
}
\examples{
# null
}
