% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_functions_fixed.R
\name{interactive_ridgeplot}
\alias{interactive_ridgeplot}
\title{Interactive Ridge Plot}
\usage{
interactive_ridgeplot(
  p,
  obj,
  features,
  group.by = "seurat_clusters",
  hw_ratio = 0.9,
  width_px = 600,
  theme_choice = "plotly_white"
)
}
\description{
Interactive Ridge Plot
}
