% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{join_layers_if_needed}
\alias{join_layers_if_needed}
\title{Helper function to join layers for Assay5 objects}
\usage{
join_layers_if_needed(SeuratObj, assay = "RNA")
}
\arguments{
\item{SeuratObj}{A Seurat object}

\item{assay}{Assay name}
}
\value{
Seurat object with joined layers
}
\description{
Helper function to join layers for Assay5 objects
}
