% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launchSeuratExplorer}
\alias{launchSeuratExplorer}
\title{Launch shiny app}
\usage{
launchSeuratExplorer()
}
\value{
In-browser Shiny Application launch
}
\description{
Launch shiny app
}
\examples{
# launchSeuratExplorer()
}
