% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launchSeuratExplorer}
\alias{launchSeuratExplorer}
\title{Launch shiny app}
\usage{
launchSeuratExplorer(
  verbose = FALSE,
  ReductionKeyWords = c("umap", "tsne", "pca"),
  SplitOptionMaxLevel = 12,
  MaxInputFileSize = 20 * 1024^3,
  DemoDataURL =
    c("https://raw.githubusercontent.com/fentouxungui/SeuratExplorerServer/refs/heads/main/inst/extdata/source-data/fly/Rds-file/G101_PC20res04.rds",
    "https://drive.usercontent.google.com/download?id=1iscfl4zyNbtjAol0bwndnBI08MDcZvYE&export=download&authuser=0&confirm=t&uuid=6352f610-9125-4551-afaa-e1eedd15701c&at=AMrWOn0YlTymrOPO6Rb0GlFtJ07o\%3A1790127440033"),
  DemoDataCacheDir = NULL,
  DemoDataMD5 = "e144cabc42684f037e60a603f74ed4f1"
)
}
\arguments{
\item{verbose}{for debug use}

\item{ReductionKeyWords}{key words used for prepare Reduction options}

\item{SplitOptionMaxLevel}{max level cutoff for prepare Split options}

\item{MaxInputFileSize}{set the limited upload file size}

\item{DemoDataURL}{one or more URLs of the demo Seurat object downloaded by
the "Download and Run a demo data" button. Multiple URLs are tried in
order until one succeeds (primary, then backup)}

\item{DemoDataCacheDir}{directory used to cache the downloaded demo data;
if NULL, a per-user cache directory is used}

\item{DemoDataMD5}{expected md5 checksum of the downloaded demo data;
set to NULL to skip the integrity check}
}
\value{
In-browser Shiny Application launch
}
\description{
used to launch the shiny app in a web browser.
}
\examples{
if(interactive()){launchSeuratExplorer()}
}
