% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{top_genes}
\alias{top_genes}
\title{各个细胞内的高表达基因}
\usage{
top_genes(SeuratObj, expr.cut = 0.01, group.by)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{expr.cut}{在单个细胞中，如果某个基因的UMI比例高于此值，则会被认为是被高表达的基因。}

\item{group.by}{细胞分群参数}
}
\value{
data frame
}
\description{
对各群里的各个单细胞，分别统计top表达的基因，即基因的reads比例大于所设定的阈值expr.cut。
}
\examples{
# top_genes(cds, expr.cut = 0.01, group.by = "orig.ident")
# top_genes(cds, expr.cut = 0.1, group.by = "res.0.4")
# top_genes(cds, expr.cut = 0.1, group.by = "RandomGroup")
}
